pyproject.toml
[tool.poetry]
name = "robo_arb"
version = "0.1.0"
description = ""
authors = ["dr.Livsy <plumbus.rise@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-binance = "^1.0.15"
asyncio = "^3.4.3"
config = "^0.5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
